notifier: move InhibitRuleToInhibitionRule to legacy_storage/v1 package
