+++ pkg/services/ngalert/notifier/legacy_storage/compat.go
@@ -6,18 +6,14 @@ import (
 	"fmt"
 	"hash/fnv"
 	"maps"
-	"strings"
 
 	alertingNotify "github.com/grafana/alerting/notify"
 	"github.com/grafana/alerting/receivers/schema"
-	"github.com/prometheus/alertmanager/config"
 	"github.com/prometheus/common/model"
-	k8svalidation "k8s.io/apimachinery/pkg/util/validation"
 
 	apimodels "github.com/grafana/grafana/pkg/services/ngalert/api/tooling/definitions"
 	"github.com/grafana/grafana/pkg/services/ngalert/models"
 	v1 "github.com/grafana/grafana/pkg/services/ngalert/notifier/legacy_storage/v1"
-	"github.com/grafana/grafana/pkg/services/sqlstore/migrations/ualert"
 )
 
 var NameToUid = models.NameToUid
@@ -280,29 +276,4 @@ func ToGroupBy(groupByStr ...string) (groupByAll bool, groupBy []model.LabelName
 		}
 	}
 	return false, groupBy
-}
-
-func InhibitRuleToInhibitionRule(name string, rule config.InhibitRule, provenance v1.Provenance) (*v1.InhibitionRule, error) {
-	if name = strings.TrimSpace(name); name == "" {
-		return nil, fmt.Errorf("inhibition rule name must not be empty")
-	}
-
-	if strings.Contains(name, ":") {
-		return nil, fmt.Errorf("inhibition rule name cannot contain invalid character ':'")
-	}
-
-	if errs := k8svalidation.IsDNS1123Subdomain(name); len(errs) > 0 {
-		return nil, fmt.Errorf("inhibition rule name must be a valid DNS subdomain: %s", strings.Join(errs, ", "))
-	}
-
-	// imported inhibition rules have purposefully long names to ensure no conflict with non-imported ones
-	if models.Provenance(provenance) != models.ProvenanceConvertedPrometheus && len(name) > ualert.UIDMaxLength {
-		return nil, fmt.Errorf("inhibition rule name is too long (exceeds %d characters)", ualert.UIDMaxLength)
-	}
-
-	return &v1.InhibitionRule{
-		Name:        name,
-		InhibitRule: rule,
-		Provenance:  provenance,
-	}, nil
 }
