+++ pkg/services/ngalert/notifier/legacy_storage/compat_test.go
@@ -1,7 +1,6 @@
 package legacy_storage
 
 import (
-	"errors"
 	"fmt"
 	"reflect"
 	"strings"
@@ -14,8 +13,6 @@ import (
 	"github.com/grafana/alerting/receivers/schema"
 	"github.com/grafana/alerting/receivers/teams"
 	webhookV0 "github.com/grafana/alerting/receivers/webhook/v0mimir1"
-	"github.com/prometheus/alertmanager/config"
-	"github.com/prometheus/alertmanager/pkg/labels"
 	"github.com/prometheus/common/model"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -185,94 +182,4 @@ func TestManagedRouteToRoute(t *testing.T) {
 	assert.Equal(t, &ri, route.RepeatInterval)
 	assert.Len(t, route.Routes, 1)
 	assert.EqualValues(t, v1.Provenance("test"), route.Provenance)
-}
-
-func Test_InhibitRuleToInhibitionRule(t *testing.T) {
-	testRule := config.InhibitRule{
-		SourceMatchers: config.Matchers{
-			{
-				Type:  labels.MatchEqual,
-				Name:  "instance",
-				Value: "alertmanager-1",
-			},
-		},
-		TargetMatchers: config.Matchers{
-			{
-				Type:  labels.MatchEqual,
-				Name:  "instance",
-				Value: "alertmanager-2",
-			},
-		},
-		Equal: []string{
-			"service",
-		},
-	}
-
-	tt := []struct {
-		name        string
-		ruleName    string
-		provenance  v1.Provenance
-		inhibitRule config.InhibitRule
-		origin      models.ResourceOrigin
-		exp         *v1.InhibitionRule
-		expErr      error
-	}{
-		{
-			name:     "fails when name is empty",
-			ruleName: "  ",
-			expErr:   errors.New("inhibition rule name must not be empty"),
-		},
-		{
-			name:     "fails when name contains ':'",
-			ruleName: "a:b",
-			expErr:   errors.New("inhibition rule name cannot contain invalid character ':'"),
-		},
-		{
-			name:     "fails when name is not a valid dns 1123 subdomain",
-			ruleName: "_some_name",
-			expErr:   errors.New("inhibition rule name must be a valid DNS subdomain: a lowercase RFC 1123 subdomain must consist of lower case alphanumeric characters, '-' or '.', and must start and end with an alphanumeric character (e.g. 'example.com', regex used for validation is '[a-z0-9]([-a-z0-9]*[a-z0-9])?(\\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*')"),
-		},
-		{
-			name:     "fails when length of non-imported rule name is over UIDMaxLength limit",
-			ruleName: "some-really-long-inhibition-rule-name-001",
-			origin:   models.ResourceOriginGrafana,
-			expErr:   errors.New("inhibition rule name is too long (exceeds 40 characters)"),
-		},
-		{
-			name:        "allows length of imported rule name to be over UIDMaxLength limit",
-			ruleName:    "some-really-long-inhibition-rule-name-001",
-			provenance:  v1.Provenance(models.ProvenanceConvertedPrometheus),
-			origin:      models.ResourceOriginImported,
-			inhibitRule: testRule,
-			exp: &v1.InhibitionRule{
-				Name:        "some-really-long-inhibition-rule-name-001",
-				InhibitRule: testRule,
-				Provenance:  v1.Provenance(models.ProvenanceConvertedPrometheus),
-			},
-		},
-		{
-			name:        "converts model correctly when all validations passes",
-			ruleName:    "inhibition-rule-1",
-			origin:      models.ResourceOriginGrafana,
-			provenance:  v1.Provenance(models.ProvenanceNone),
-			inhibitRule: testRule,
-			exp: &v1.InhibitionRule{
-				Name:        "inhibition-rule-1",
-				InhibitRule: testRule,
-				Provenance:  v1.Provenance(models.ProvenanceNone),
-			},
-		},
-	}
-
-	for _, tc := range tt {
-		t.Run(tc.name, func(t *testing.T) {
-			got, gotErr := InhibitRuleToInhibitionRule(tc.ruleName, tc.inhibitRule, tc.provenance)
-			if tc.expErr != nil {
-				require.EqualError(t, gotErr, tc.expErr.Error())
-			} else {
-				require.Nil(t, gotErr)
-			}
-			require.Equal(t, tc.exp, got)
-		})
-	}
 }
