+++ pkg/services/ngalert/notifier/legacy_storage/v1/compat_test.go
@@ -2,13 +2,17 @@ package v1
 
 import (
 	"encoding/json"
+	"errors"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/grafana/alerting/definition"
+	"github.com/prometheus/alertmanager/config"
 	"github.com/prometheus/alertmanager/pkg/labels"
 	"github.com/stretchr/testify/require"
+
+	"github.com/grafana/grafana/pkg/services/ngalert/models"
 )
 
 func TestRoundTripConversion(t *testing.T) {
@@ -198,4 +202,94 @@ func TestRoundTripConversion(t *testing.T) {
 
 	require.JSONEq(t, configJSON, string(convertedJSON),
 		"Round-trip conversion should be lossless")
+}
+
+func Test_InhibitRuleToInhibitionRule(t *testing.T) {
+	testRule := config.InhibitRule{
+		SourceMatchers: config.Matchers{
+			{
+				Type:  labels.MatchEqual,
+				Name:  "instance",
+				Value: "alertmanager-1",
+			},
+		},
+		TargetMatchers: config.Matchers{
+			{
+				Type:  labels.MatchEqual,
+				Name:  "instance",
+				Value: "alertmanager-2",
+			},
+		},
+		Equal: []string{
+			"service",
+		},
+	}
+
+	tt := []struct {
+		name        string
+		ruleName    string
+		provenance  Provenance
+		inhibitRule config.InhibitRule
+		origin      models.ResourceOrigin
+		exp         *InhibitionRule
+		expErr      error
+	}{
+		{
+			name:     "fails when name is empty",
+			ruleName: "  ",
+			expErr:   errors.New("inhibition rule name must not be empty"),
+		},
+		{
+			name:     "fails when name contains ':'",
+			ruleName: "a:b",
+			expErr:   errors.New("inhibition rule name cannot contain invalid character ':'"),
+		},
+		{
+			name:     "fails when name is not a valid dns 1123 subdomain",
+			ruleName: "_some_name",
+			expErr:   errors.New("inhibition rule name must be a valid DNS subdomain: a lowercase RFC 1123 subdomain must consist of lower case alphanumeric characters, '-' or '.', and must start and end with an alphanumeric character (e.g. 'example.com', regex used for validation is '[a-z0-9]([-a-z0-9]*[a-z0-9])?(\\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*')"),
+		},
+		{
+			name:     "fails when length of non-imported rule name is over UIDMaxLength limit",
+			ruleName: "some-really-long-inhibition-rule-name-001",
+			origin:   models.ResourceOriginGrafana,
+			expErr:   errors.New("inhibition rule name is too long (exceeds 40 characters)"),
+		},
+		{
+			name:        "allows length of imported rule name to be over UIDMaxLength limit",
+			ruleName:    "some-really-long-inhibition-rule-name-001",
+			provenance:  Provenance(models.ProvenanceConvertedPrometheus),
+			origin:      models.ResourceOriginImported,
+			inhibitRule: testRule,
+			exp: &InhibitionRule{
+				Name:        "some-really-long-inhibition-rule-name-001",
+				InhibitRule: testRule,
+				Provenance:  Provenance(models.ProvenanceConvertedPrometheus),
+			},
+		},
+		{
+			name:        "converts model correctly when all validations passes",
+			ruleName:    "inhibition-rule-1",
+			origin:      models.ResourceOriginGrafana,
+			provenance:  Provenance(models.ProvenanceNone),
+			inhibitRule: testRule,
+			exp: &InhibitionRule{
+				Name:        "inhibition-rule-1",
+				InhibitRule: testRule,
+				Provenance:  Provenance(models.ProvenanceNone),
+			},
+		},
+	}
+
+	for _, tc := range tt {
+		t.Run(tc.name, func(t *testing.T) {
+			got, gotErr := InhibitRuleToInhibitionRule(tc.ruleName, tc.inhibitRule, tc.provenance)
+			if tc.expErr != nil {
+				require.EqualError(t, gotErr, tc.expErr.Error())
+			} else {
+				require.Nil(t, gotErr)
+			}
+			require.Equal(t, tc.exp, got)
+		})
+	}
 }
